Utils implementado, Tabla de transposicion v1.0

+++ src/edu/upc/epsevg/prop/checkers/players/javakillers/utils/Node.java
@@ -9,9 +9,9 @@ package edu.upc.epsevg.prop.checkers.players.javakillers.utils;
  * @author ivanr
  */
 public class Node {
-    public int  eval;
+    public double  eval;
 
-    public Node( int eval) {
+    public Node( double eval) {
        
         this.eval = eval;
     }
